Extract linear-gradient
